+++ HISTORY.md
@@ -3,6 +3,7 @@
 ## in development:
 * **New Features:**
     * new "rgba" option in "(aspect) AT (location)" sensing reporter, returns a 4-item list of values from 0-255 (same is pixels from a costume)
+    * "temporary?" attribute is now programmatically settable (in the SET->my... block)
 * **Notable Changes:**
 * **Notable Fixes:**
 * **Translation Updates:**
@@ -10,6 +11,7 @@
 ### 2019-08-07
 * new dev version
 * blocks, threads: added "rgba" option to (aspect) AT (location) reporter in the sensing category
+* blocks, threads: made "temporary?" attribute for clones settable (in the SET->my... block)
 
 ## v5.0.9
 * **New Feature:**