+++ HISTORY.md
@@ -3,7 +3,7 @@
 ## in development:
 * **New Features:**
     * new "paste on" block in the pen category, prints a sprite onto another one
-    * new "rgba" option in "(aspect) AT (location)" sensing reporter, returns a 4-item list of values from 0-255 (same is pixels from a costume)
+    * new "r-g-b-a" option in "(aspect) AT (location)" sensing reporter, returns a 4-item list of values from 0-255 (same is pixels from a costume)
     * "temporary?" attribute is now programmatically settable (in the SET->my... block)
 * **Notable Changes:**
     * generated costumes that are not in the wardrobe are not made persistent in the project (saved & restored)
@@ -15,7 +15,7 @@
 
 ### 2019-08-07
 * new dev version
-* blocks, threads: added "rgba" option to (aspect) AT (location) reporter in the sensing category
+* blocks, threads: added "r-g-b-a" option to (aspect) AT (location) reporter in the sensing category
 * blocks, threads: made "temporary?" attribute for clones settable (in the SET->my... block)
 * objects: made the "paste on" block avaible in the "pen" palette
 * fix for tainted audio context when the user interacts with an auto-playing project